frmbase/pyproject.toml
[tool.poetry]
name = "frmbase"
version = "0.1.0"
description = ""
authors = ["Fergal <fergal.mullally@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
colorama= "*"
numpy = "*"
gitpython = "*"
scipy = "*"
pandas = "*"
tqdm = "*"
ipdb = "*"


[tool.poetry.dev-dependencies]
frmdev = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
